stick-data-collection/requirements.txt
matplotlib==3.8.1
pyliblzfse==0.4.1
opencv-python==4.8.1.78
numpy-stl==3.1.0
numpy-quaternion==2022.4.3
pyfqmr==0.2.0
open3d==0.16.0
torch==2.1.1
torchvision==0.16.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0